WASM compile

+++ src/neuropia.cpp
@@ -981,7 +981,7 @@ Sizes Layer::sizes() const {
 size_t Layer::consumption(bool cumulative) const {
     const auto c = sizeof(this) 
     + m_outBuffer.size() * sizeof(decltype(m_outBuffer)::value_type)
-    + std::accumulate(m_neurons.begin(), m_neurons.end(), 0U, [](const auto& a, const auto& n ) {
+    + std::accumulate(m_neurons.begin(), m_neurons.end(), static_cast<size_t>(0U), [](const auto& a, const auto& n ) {
         return a + n.consumption();
     });
